split settings code to more granular files and change how NotificationsHandler component is plugged into main React tree
