+++ src/notifications/settings/index.ts
@@ -1,0 +1,30 @@
+export {
+  NotificationTopic,
+  NotificationRelationship,
+  WALLET_GROUPS_STORAGE_KEY,
+  WALLET_TOPICS_STORAGE_KEY,
+} from './constants';
+export {
+  addDefaultNotificationGroupSettings,
+  addDefaultNotificationSettingsForWallet,
+} from './defaults';
+export {
+  useNotificationSettings,
+  useAllNotificationSettingsFromStorage,
+  useWalletGroupNotificationSettings,
+} from './hooks';
+export {
+  removeNotificationSettingsForWallet,
+  toggleGroupNotifications,
+  toggleTopicForWallet,
+} from './settings';
+export type {
+  GroupSettings,
+  NotificationRelationshipType,
+  NotificationTopicType,
+  WalletNotificationSettings,
+} from './types';
+export {
+  notificationSettingsStorage,
+  updateSettingsForWallets,
+} from './storage';
